backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17
opencv-python-headless==4.10.0.84
Pillow==11.0.0
supabase==2.10.0
python-dotenv==1.0.1
cairosvg>=2.7.0
